Rename test package: aegus -> hotbot
