+++ src/test/kotlin/me/aberrantfox/hotbot/extensions/InviteMatchTest.kt
@@ -1,6 +1,5 @@
-package me.aberrantfox.aegus.extenstions
+package me.aberrantfox.hotbot.extensions
 
-import me.aberrantfox.hotbot.extensions.containsInvite
 import net.dv8tion.jda.core.MessageBuilder
 import org.junit.Assert
 import org.junit.Test
